fix types

+++ pmwdlib/jax_pmwd_common.h
@@ -12,7 +12,7 @@ struct PmwdDescriptor {
   T ptcl_spacing;
   T cell_size;
   T offset[3];
-  int64_t stride[3];
+  uint32_t stride[3];
 };
 
 }
